removed comment from model.py

+++ ml_model/model.py
@@ -21,47 +21,17 @@ model.eval()
 model = model.float()
 model = model.to('cpu')
 
-# # Preprocess the image
-# im = cv2.resize(im, (224, 224))
-# im = cv2.cvtColor(im, cv2.COLOR_BGR2RGB)
-# im = im.transpose((2, 0, 1))  # Change data layout from HWC to CHW
-# im = im / 255.0  # Normalize to [0, 1]
-# im = torch.tensor(im, dtype=torch.float32)
-# im = im.unsqueeze(0)  # Add batch dimension
-# im = im.half()  # Convert to half precision
-# # im = im.to('cuda')  # Move to GPU
-
-
-# # # Run inference
-# # with torch.no_grad():
-# output = model(im)
-# # print(output)
-
-
-# # Post-process the output
-# _, predicted = torch.max(output, 1)
-# # print(predicted)
-
-# # Print the predicted class
-# print(f'Predicted class: {predicted.item()}')
-
-# # Get the corresponding label
-# predicted_label = labels[predicted.item()]
-# print(f"Predicted label: {predicted_label}")
-
-
 def get_prediction(img_path):
     im = cv2.imread(img_path)
 
     # Preprocess the image
     im = cv2.resize(im, (224, 224))
     im = cv2.cvtColor(im, cv2.COLOR_BGR2RGB)
-    im = im.transpose((2, 0, 1))  # Change data layout from HWC to CHW
-    im = im / 255.0  # Normalize to [0, 1]
+    im = im.transpose((2, 0, 1))
+    im = im / 255.0 
     im = torch.tensor(im, dtype=torch.float32)
-    im = im.unsqueeze(0)  # Add batch dimension
-    im = im.float()  # Convert to half precision
-    # im = im.to('cuda')  # Move to GPU
+    im = im.unsqueeze(0)
+    im = im.float()
 
     output = model(im)
